Improve compatibility by using a tighter mixin Mixin applied to both sides, removing redundant code.
