+++ common/src/main/resources/fishingreal-common.mixins.json
@@ -1,10 +1,10 @@
 {
   "required": true,
   "minVersion": "0.8",
-  "package": "koala.fishingreal.fabric.mixin",
+  "package": "koala.fishingreal.mixin",
   "compatibilityLevel": "JAVA_17",
   "mixins": [
-    "FishingHookMixin"
+	  "FishingHookMixin"
   ],
   "client": [
   ],
