+++ forge/src/main/java/koala/fishingreal/forge/FishingRealForge.java
@@ -9,21 +9,10 @@ import net.minecraftforge.fml.common.Mod;
 @Mod(FishingReal.MOD_ID)
 public class FishingRealForge {
     public FishingRealForge() {
-        FishingReal.init();
         MinecraftForge.EVENT_BUS.addListener(FishingRealForge::onServerReloadListeners);
-        MinecraftForge.EVENT_BUS.addListener(FishingRealForge::itemFished);
     }
 
     public static void onServerReloadListeners(AddReloadListenerEvent event) {
         FishingReal.onRegisterReloadListeners((id, listener) -> event.addListener(listener));
-    }
-
-    public static void itemFished(ItemFishedEvent event) {
-        if (FishingReal.retrieve(event.getEntity(), event.getDrops(), event.getHookEntity())) {
-            if (!event.getEntity().isCreative()) {
-                event.damageRodBy(1);
-            }
-            event.setCanceled(true);
-        }
     }
 }
